Improve handling of defs, add cache
Each time we load a new fragment, we see if we can find any `<defs>`
element in the fragment. If we do, we copy them into the global defs
element.

This also implements a cache, so we don't make ajax requests for every
`load` call. Instead, we clone the result and store it in the cache.
Each subsequent time it is requested, we clone it again.

We don't add defs a second time if we're loading from cache.