Add hover events, add rich view object setters
This adds "view.hover" and "view.hover.enter, view.hover.exit" events,
with associated event targets.

This also makes it easier to set properties on all objects in the view,
and the re-set them to initial values. We can also set properties in
the style {propName: "*multiple"}, with the four basic math functions,
to allow relative changes of properties.