Agent can find, move from nearest point along path
This adds a new location `at` option, `nearest`. Given an existing
agent location, `nearest` will find the nearest point on a path.

This can be used in the `move_to` and `follow` tasks.

Also add the option `direction : "random"` for the `follow` task.

Update the cb-melanocyte model with the ability to click in the
intercell region and add hormone.

[#155239645]